+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 * Fixed opening the generated `*.xlsx` files in MS Office 2007 Excel.
 
+* Added `align` attribute.
+
 1.0.0 / 17.06.2021
 ==================